separate DataTable from DataTableView
