+++ apps/src/storage/dataBrowser/DataTable.jsx
@@ -1,9 +1,7 @@
 /**
- * @overview Component for editing a data table.
+ * @overview Component for displaying a data table.
  */
-import TableControls from './TableControls';
 import AddTableRow from './AddTableRow';
-import {DataView} from '../constants';
 import EditTableRow from './EditTableRow';
 import ColumnHeader from './ColumnHeader';
 import FirebaseStorage from '../firebaseStorage';
@@ -11,14 +9,12 @@ import FontAwesome from '../../templates/FontAwesome';
 import PropTypes from 'prop-types';
 import Radium from 'radium';
 import React from 'react';
-import {changeView, showWarning} from '../redux/data';
+import {showWarning} from '../redux/data';
 import * as dataStyles from './dataStyles';
 import color from '../../util/color';
 import {connect} from 'react-redux';
-import {getColumnNamesFromRecords} from '../firebaseMetadata';
 import PaginationWrapper from '../../templates/PaginationWrapper';
 import msg from '@cdo/locale';
-import experiments from '../../util/experiments';
 
 const MIN_TABLE_WIDTH = 600;
 const MAX_ROWS_PER_PAGE = 500;
@@ -30,26 +26,12 @@ const styles = {
       width: 19
     }
   ],
-  container: {
-    flexDirection: 'column',
-    height: '100%',
-    minWidth: MIN_TABLE_WIDTH,
-    maxWidth: '100%',
-    paddingLeft: experiments.isEnabled(experiments.APPLAB_DATASETS)
-      ? '8px'
-      : '0px'
-  },
   table: {
-    minWidth: MIN_TABLE_WIDTH
-  },
-  tableWrapper: {
-    flexGrow: 1,
+    minWidth: MIN_TABLE_WIDTH,
     overflow: 'scroll'
   },
   pagination: {
-    float: 'right',
-    display: 'inline',
-    marginTop: 10
+    marginBottom: 20
   },
   plusIcon: {
     alignItems: 'center',
@@ -69,12 +51,12 @@ const INITIAL_STATE = {
   pendingAdd: false,
   // The old name of the column currently being renamed or deleted.
   pendingColumn: null,
-  showDebugView: false,
   currentPage: 0
 };
 
 class DataTable extends React.Component {
   static propTypes = {
+    getColumnNames: PropTypes.func.isRequired,
     // from redux state
     tableColumns: PropTypes.arrayOf(PropTypes.string).isRequired,
     tableName: PropTypes.string.isRequired,
@@ -84,11 +66,9 @@ class DataTable extends React.Component {
     // https://firebase.googleblog.com/2014/04/best-practices-arrays-in-firebase.html
     tableRecords: PropTypes.oneOfType([PropTypes.object, PropTypes.array])
       .isRequired,
-    view: PropTypes.oneOf(Object.keys(DataView)),
 
     // from redux dispatch
-    onShowWarning: PropTypes.func.isRequired,
-    onViewChange: PropTypes.func.isRequired
+    onShowWarning: PropTypes.func.isRequired
   };
 
   state = {...INITIAL_STATE};
@@ -178,25 +158,8 @@ class DataTable extends React.Component {
     });
   };
 
-  /**
-   * @param {Array} records Array of JSON-encoded records.
-   * @param {string} columns Array of column names.
-   */
-  getColumnNames(records, columns) {
-    // Make sure 'id' is the first column.
-    const columnNames = getColumnNamesFromRecords(records);
-
-    columns.forEach(columnName => {
-      if (columnNames.indexOf(columnName) === -1) {
-        columnNames.push(columnName);
-      }
-    });
-
-    return columnNames;
-  }
-
   getNextColumnName() {
-    const names = this.getColumnNames(
+    const names = this.props.getColumnNames(
       this.props.tableRecords,
       this.props.tableColumns
     );
@@ -206,46 +169,6 @@ class DataTable extends React.Component {
     }
     return `column${i}`;
   }
-
-  importCsv = (csvData, onComplete) => {
-    FirebaseStorage.importCsv(
-      this.props.tableName,
-      csvData,
-      () => {
-        this.setState(INITIAL_STATE);
-        onComplete();
-      },
-      msg => {
-        if (String(msg).includes('data is too large')) {
-          this.props.onShowWarning(msg);
-        } else {
-          console.warn(msg);
-        }
-        onComplete();
-      }
-    );
-  };
-
-  exportCsv = () => {
-    const tableName = encodeURIComponent(this.props.tableName);
-    location.href = `/v3/export-firebase-tables/${
-      Applab.channelId
-    }/${tableName}`;
-  };
-
-  /** Delete all rows, but preserve the columns. */
-  clearTable = () => {
-    FirebaseStorage.clearTable(
-      this.props.tableName,
-      () => {},
-      msg => console.warn(msg)
-    );
-  };
-
-  toggleDebugView = () => {
-    const showDebugView = !this.state.showDebugView;
-    this.setState({showDebugView});
-  };
 
   coerceColumn = (columnName, columnType) => {
     this.setState({
@@ -270,16 +193,6 @@ class DataTable extends React.Component {
     }, 0);
   };
 
-  getTableJson() {
-    const records = [];
-    // Cast Array to Object
-    const tableRecords = Object.assign({}, this.props.tableRecords);
-    for (const id in tableRecords) {
-      records.push(JSON.parse(tableRecords[id]));
-    }
-    return JSON.stringify(records, null, 2);
-  }
-
   onChangePageNumber = number => {
     this.setState({currentPage: number - 1});
   };
@@ -295,7 +208,7 @@ class DataTable extends React.Component {
   }
 
   render() {
-    let columnNames = this.getColumnNames(
+    let columnNames = this.props.getColumnNames(
       this.props.tableRecords,
       this.props.tableColumns
     );
@@ -313,59 +226,8 @@ class DataTable extends React.Component {
       columnNames.push(editingColumn);
     }
 
-    const visible = DataView.TABLE === this.props.view;
-    const containerStyle = [
-      styles.container,
-      {
-        display: visible ? '' : 'none'
-      }
-    ];
-    const tableDataStyle = [
-      styles.table,
-      {
-        display: this.state.showDebugView ? 'none' : ''
-      }
-    ];
-    const debugDataStyle = [
-      dataStyles.debugData,
-      {
-        display: this.state.showDebugView ? '' : 'none'
-      }
-    ];
     return (
-      <div id="dataTable" style={containerStyle} className="inline-flex">
-        <div style={dataStyles.viewHeader}>
-          <span style={dataStyles.backLink}>
-            <a
-              id="tableBackToOverview"
-              style={dataStyles.link}
-              onClick={() => this.props.onViewChange(DataView.OVERVIEW)}
-            >
-              <FontAwesome icon="arrow-circle-left" />
-              &nbsp;Back to data
-            </a>
-          </span>
-          <span style={dataStyles.debugLink}>
-            <a
-              id="uitest-tableDebugLink"
-              style={dataStyles.link}
-              onClick={this.toggleDebugView}
-            >
-              {this.state.showDebugView ? 'Table view' : 'Debug view'}
-            </a>
-          </span>
-        </div>
-
-        <TableControls
-          columns={columnNames}
-          clearTable={this.clearTable}
-          importCsv={this.importCsv}
-          exportCsv={this.exportCsv}
-          tableName={this.props.tableName}
-        />
-
-        <div style={debugDataStyle}>{this.getTableJson()}</div>
-
+      <div>
         <div style={styles.pagination}>
           <PaginationWrapper
             totalPages={numPages}
@@ -374,9 +236,8 @@ class DataTable extends React.Component {
             label={msg.paginationLabel()}
           />
         </div>
-
-        <div style={styles.tableWrapper}>
-          <table style={tableDataStyle}>
+        <div>
+          <table>
             <tbody>
               <tr>
                 {columnNames.map(columnName => (
@@ -439,7 +300,6 @@ class DataTable extends React.Component {
 
 export default connect(
   state => ({
-    view: state.data.view,
     tableColumns: state.data.tableColumns || [],
     tableRecords: state.data.tableRecords || {},
     tableName: state.data.tableName || ''
@@ -447,9 +307,6 @@ export default connect(
   dispatch => ({
     onShowWarning(warningMsg, warningTitle) {
       dispatch(showWarning(warningMsg, warningTitle));
-    },
-    onViewChange(view) {
-      dispatch(changeView(view));
     }
   })
 )(Radium(DataTable));
